server/planme_backend/requirements.txt
flask
flask-cors
flask-restful
bcrypt
psycopg2
pyjwt